Refactor to use frontend packages
I believe this covers all working functionality 1:1 except for the TOCWidget that depends on the endpoints plugin that we need to make/port/package
