+++ package.json
@@ -10,46 +10,56 @@
     "testDebug": "node --inspect-brk node_modules/.bin/jest --runInBand"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.26",
-    "@fortawesome/free-solid-svg-icons": "^5.12.0",
-    "@fortawesome/vue-fontawesome": "^0.1.9",
-    "@scaife-viewer/scaife-widgets": "git://github.com/scaife-viewer/scaife-widgets#508b48f9aaad6c452916b295d8e87de49d5f9915",
-    "apollo-boost": "^0.4.7",
-    "core-js": "^3.5.0",
-    "graphql": "^14.4.2",
-    "graphql-tag": "^2.10.1",
-    "lodash.clonedeep": "^4.5.0",
-    "node-sass": "^4.13.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.30",
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
+    "@fortawesome/vue-fontawesome": "^0.1.10",
+    "@scaife-viewer/skeleton": "^0.0.4",
+    "@scaife-viewer/store": "^0.0.3",
+    "@scaife-viewer/widget-library": "^0.0.3",
+    "@scaife-viewer/widget-metadata": "^0.0.3",
+    "@scaife-viewer/widget-new-alexandria": "^0.0.3",
+    "@scaife-viewer/widget-passage-ancestors": "^0.0.3",
+    "@scaife-viewer/widget-passage-children": "^0.0.3",
+    "@scaife-viewer/widget-passage-reference": "^0.0.6",
+    "@scaife-viewer/widget-reader": "^0.0.3",
+    "@scaife-viewer/widget-text-size": "^0.0.4",
+    "@scaife-viewer/widget-text-width": "^0.0.4",
+    "@scaife-viewer/widget-toc": "^0.0.3",
+    "@scaife-viewer/widget-word-list": "^0.0.3",
+    "apollo-boost": "^0.4.9",
+    "core-js": "^3.6.5",
+    "graphql-tag": "^2.11.0",
+    "node-sass": "^4.14.1",
     "path": "^0.12.7",
-    "sass-loader": "^8.0.0",
-    "scaife-skeleton": "^0.7.0",
+    "sass-loader": "^9.0.3",
     "vue": "^2.6.11",
-    "vue-router": "^3.0.3",
-    "vuex": "^3.0.1",
-    "webpack": "^4.41.3"
+    "vue-apollo": "^3.0.4",
+    "vue-router": "^3.4.3",
+    "vuex": "^3.5.1",
+    "vuex-router-sync": "^5.0.0",
+    "webpack": "^4.44.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.5",
-    "@babel/preset-env": "^7.7.6",
-    "@vue/cli-plugin-babel": "^4.1.1",
-    "@vue/cli-plugin-eslint": "^4.1.1",
-    "@vue/cli-plugin-unit-jest": "^4.2.2",
-    "@vue/cli-service": "^4.1.1",
-    "@vue/eslint-config-airbnb": "^5.0.0",
+    "@babel/core": "^7.11.1",
+    "@babel/preset-env": "^7.11.0",
+    "@vue/cli-plugin-babel": "^4.5.4",
+    "@vue/cli-plugin-eslint": "^4.5.4",
+    "@vue/cli-plugin-unit-jest": "^4.5.4",
+    "@vue/cli-service": "^4.5.4",
+    "@vue/eslint-config-airbnb": "^5.1.0",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/server-test-utils": "^1.0.0-beta.30",
-    "@vue/test-utils": "^1.0.0-beta.30",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.3",
-    "babel-jest": "^24.9.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-vue": "^6.0.1",
-    "jest": "^24.9.0",
-    "prettier": "^1.19.1",
-    "style-resources-loader": "^1.3.2",
+    "@vue/server-test-utils": "^1.0.4",
+    "@vue/test-utils": "^1.0.4",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^26.3.0",
+    "eslint": "^7.7.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-vue": "^6.2.2",
+    "jest": "^26.4.0",
+    "prettier": "^2.0.5",
+    "style-resources-loader": "^1.3.3",
     "vue-cli-plugin-style-resources-loader": "^0.1.4",
-    "vue-jest": "^3.0.5",
+    "vue-jest": "^3.0.6",
     "vue-server-renderer": "2.6.11",
     "vue-template-compiler": "^2.6.11"
   },
