+++ src/App.vue
@@ -15,7 +15,6 @@
   };
 </script>
 
-<style src="@scaife-viewer/scaife-widgets/dist/scaife-widgets.css"></style>
 <style lang="scss">
   html,
   body {
