--- src/constants.js
@@ -1,5 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const FETCH_METADATA = 'fetchMetadata';
-export const UPDATE_METADATA = 'updateMetadata';
-export const FETCH_LIBRARY = 'fetchLibrary';
-export const SET_PASSAGE = 'setPassage';
