--- src/mutations.js
@@ -1,13 +1,0 @@
-import { FETCH_METADATA, FETCH_LIBRARY, SET_PASSAGE } from '@/constants';
-
-export default {
-  [FETCH_METADATA]: (state, metadata) => {
-    state.metadata = metadata;
-  },
-  [FETCH_LIBRARY]: (state, libraryTree) => {
-    state.libraryTree = libraryTree;
-  },
-  [SET_PASSAGE]: (state, urn) => {
-    state.passage = urn;
-  },
-};
