--- src/reader/actions.js
@@ -1,1 +1,0 @@
-export default {};
