--- src/reader/constants.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const MODULE_NS = 'scaifeReader';
