--- src/reader/getters.js
@@ -1,6 +1,0 @@
-export default {
-  workTitle: (state, getters, rootState, rootGetters) => rootGetters.workTitle,
-  firstPassageUrn: (state, getters, rootState, rootGetters) =>
-    rootGetters.firstPassageUrn,
-  passage: (state, getters, rootState, rootGetters) => rootGetters.passage,
-};
