--- src/reader/mutations.js
@@ -1,1 +1,0 @@
-export default {};
