--- src/reader/state.js
@@ -1,1 +1,0 @@
-export default {};
