--- src/state.js
@@ -1,5 +1,0 @@
-export default {
-  metadata: null,
-  passage: null,
-  libraryTree: null,
-};
