+++ src/views/TOCsView.vue
@@ -3,16 +3,14 @@
 </template>
 
 <script>
-  import { TOCWidget, HelloWorld } from '@scaife-viewer/scaife-widgets';
-
   export default {
     name: 'TOCsView',
     computed: {
       mainWidget() {
-        return TOCWidget;
+        return null;
       },
       rightWidgets() {
-        return [HelloWorld];
+        return [];
       },
     },
   };
